add theme class
