+++ resources/js/logged_in/components/CreateList/CreateList.jsx
@@ -7,7 +7,8 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
-
+import { withStyles } from '@material-ui/core/styles';
+import styles from "./style";
 
 function CreateList(props) {
 
@@ -70,4 +71,4 @@ function CreateList(props) {
     );
 }
 
-export default CreateList;
+export default (withStyles(styles, { withTheme: true })(CreateList));
