+++ resources/js/logged_in/components/CreateList/index.js
@@ -1,0 +1,3 @@
+import page from './CreateList';
+
+export default page;
