--- resources/js/logged_in/components/Footer.js
@@ -1,19 +1,0 @@
-import React, { Component } from "react";
-
-class Footer extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-        }
-    }
-
-    render() {
-        return (
-            <div className="footer">
-                footer
-            </div>
-        );
-    }
-}
-
-export default Footer;
