+++ resources/js/logged_in/components/ListsPredefined/index.js
@@ -1,0 +1,3 @@
+import page from './ListsPredefined';
+
+export default page;
