+++ resources/js/logged_in/components/ListsUser/index.js
@@ -1,0 +1,3 @@
+import page from './ListsUser';
+
+export default page;
