+++ resources/js/logged_in/components/MainPage/index.js
@@ -1,0 +1,3 @@
+import page from './MainPage';
+
+export default page;
