+++ resources/js/logged_in/components/MainPage/style.js
@@ -1,0 +1,14 @@
+const styles = (theme) => ({
+    main: {
+      marginLeft: theme.spacing(28),
+      transition: theme.transitions.create(["width", "margin"], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      [theme.breakpoints.down("xs")]: {
+        marginLeft: 0,
+      },
+    },
+});
+
+export default styles;
