+++ resources/js/logged_in/components/TaskItem/index.js
@@ -1,0 +1,3 @@
+import page from './TaskItem';
+
+export default page;
