+++ resources/js/logged_in/components/TaskItem/style.js
@@ -1,0 +1,15 @@
+const styles = (theme) => ({
+  
+  brandText: {
+    fontFamily: "'Baloo Bhaijaan', cursive",
+    fontWeight: 400,
+  },
+  
+  permanentDrawerListItem: {
+    justifyContent: "center",
+    paddingTop: theme.spacing(2),
+    paddingBottom: theme.spacing(2),
+  },
+});
+
+export default styles;
