+++ resources/js/logged_in/components/TaskPredefined/TaskPredefined.jsx
@@ -1,6 +1,8 @@
 import React, { Component } from "react";
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import StarBorder from '@material-ui/icons/StarBorder';
+import { withStyles } from '@material-ui/core/styles';
+import styles from "./style";
 class TaskPredefined extends Component {
     constructor(props) {
         super(props);
@@ -17,4 +19,4 @@ class TaskPredefined extends Component {
     }
 }
 
-export default TaskPredefined;
+export default (withStyles(styles, { withTheme: true })(TaskPredefined));
