+++ resources/js/logged_in/components/TaskPredefined/index.js
@@ -1,0 +1,3 @@
+import page from './TaskPredefined';
+
+export default page;
