+++ resources/js/logged_in/components/TaskPredefined/style.js
@@ -1,0 +1,6 @@
+const styles = (theme) => ({
+  
+ 
+});
+
+export default styles;
