+++ resources/js/logged_in/components/Tasks/index.js
@@ -1,0 +1,3 @@
+import page from './Tasks';
+
+export default page;
