+++ resources/js/logged_in/components/Tasks/style.js
@@ -1,0 +1,11 @@
+const styles = (theme) => ({
+  
+  brandText: {
+    fontFamily: "'Baloo Bhaijaan', cursive",
+    fontWeight: 400,
+  },
+  
+  
+});
+
+export default styles;
