+++ resources/js/logged_in/components/navigation/MessageListItem/index.js
@@ -1,0 +1,3 @@
+import page from './MessageListItem';
+
+export default page;
