+++ resources/js/logged_in/components/navigation/MessageListItem/style.js
@@ -1,0 +1,5 @@
+const styles = (theme) => ({
+  
+});
+
+export default styles;
