+++ resources/js/logged_in/components/navigation/NavBar/index.js
@@ -1,0 +1,3 @@
+import page from './NavBar';
+
+export default page;
