+++ resources/js/logged_in/components/navigation/NavBar/style.js
@@ -1,0 +1,96 @@
+const styles = (theme) => ({
+  appBar: {
+    boxShadow: theme.shadows[6],
+    backgroundColor: theme.palette.common.white,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+    [theme.breakpoints.down("xs")]: {
+      width: "100%",
+      marginLeft: 0,
+    },
+  },
+  appBarToolbar: {
+    display: "flex",
+    justifyContent: "space-between",
+    paddingLeft: theme.spacing(1),
+    paddingRight: theme.spacing(1),
+    [theme.breakpoints.up("sm")]: {
+      paddingLeft: theme.spacing(2),
+      paddingRight: theme.spacing(2),
+    },
+    [theme.breakpoints.up("md")]: {
+      paddingLeft: theme.spacing(3),
+      paddingRight: theme.spacing(3),
+    },
+    [theme.breakpoints.up("lg")]: {
+      paddingLeft: theme.spacing(4),
+      paddingRight: theme.spacing(4),
+    },
+  },
+  accountAvatar: {
+    backgroundColor: theme.palette.secondary.main,
+    height: 24,
+    width: 24,
+    marginLeft: theme.spacing(2),
+    marginRight: theme.spacing(2),
+    [theme.breakpoints.down("xs")]: {
+      marginLeft: theme.spacing(1.5),
+      marginRight: theme.spacing(1.5),
+    },
+  },
+  drawerPaper: {
+    height: "100%vh",
+    whiteSpace: "nowrap",
+    border: 0,
+    width: theme.spacing(7),
+    overflowX: "hidden",
+    marginTop: theme.spacing(8),
+    [theme.breakpoints.up("sm")]: {
+      width: theme.spacing(27),
+    },
+    backgroundColor: theme.palette.common.black,
+  },
+  smBordered: {
+    [theme.breakpoints.down("xs")]: {
+      borderRadius: "50% !important",
+    },
+  },
+  menuLink: {
+    textDecoration: "none",
+    color: theme.palette.text.primary,
+  },
+  iconListItem: {
+    width: "auto",
+    borderRadius: theme.shape.borderRadius,
+    paddingTop: 11,
+    paddingBottom: 11,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
+  },
+  textPrimary: {
+    color: theme.palette.primary.main,
+  },
+  mobileItemSelected: {
+    backgroundColor: `${theme.palette.primary.main} !important`,
+  },
+  brandText: {
+    fontFamily: "'Baloo Bhaijaan', cursive",
+    fontWeight: 400,
+  },
+  username: {
+    paddingLeft: 0,
+    paddingRight: theme.spacing(2),
+  },
+  justifyCenter: {
+    justifyContent: "center",
+  },
+  permanentDrawerListItem: {
+    justifyContent: "center",
+    paddingTop: theme.spacing(2),
+    paddingBottom: theme.spacing(2),
+  },
+});
+
+export default styles;
