--- resources/js/logged_in/components/navigation/SideDrawer.js
@@ -1,57 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import {
-  Drawer,
-  IconButton,
-  Toolbar,
-  Divider,
-  Typography,
-  Box,
-  withStyles
-} from "@material-ui/core";
-import CloseIcon from "@material-ui/icons/Close";
-
-const drawerWidth = 240;
-
-const styles = {
-  toolbar: {
-    minWidth: drawerWidth
-  }
-};
-
-function SideDrawer(props) {
-  const { classes, onClose, open } = props;
-  return (
-    <Drawer anchor="right" open={open} variant="temporary" onClose={onClose}>
-      <Toolbar disableGutters className={classes.toolbar}>
-        <Box
-          pl={3}
-          pr={3}
-          display="flex"
-          justifyContent="space-between"
-          width="100%"
-          alignItems="center"
-        >
-          <Typography variant="h6">A Sidedrawer</Typography>
-          <IconButton
-            onClick={onClose}
-            color="primary"
-            aria-label="Close Sidedrawer"
-          >
-            <CloseIcon />
-          </IconButton>
-          
-        </Box>
-      </Toolbar>
-      <Divider />
-    </Drawer>
-  );
-}
-
-// SideDrawer.propTypes = {
-//   classes: PropTypes.object.isRequired,
-//   open: PropTypes.bool.isRequired,
-//   onClose: PropTypes.func.isRequired
-// };
-
-export default withStyles(styles)(SideDrawer);
